setup as a gem
