+++ Gemfile
@@ -1,11 +1,13 @@
 source 'https://rubygems.org'
 
-gem 'rb-appscript'
-gem 'json'
-gem 'trollop'
-gem 'jira-ruby', :git => 'https://github.com/sumoheavy/jira-ruby.git'
-gem 'highline'
+gemspec :path => '.'
 
-gem 'hashie'
-gem 'pry'
+# gem 'rb-appscript'
+# gem 'json'
+# gem 'trollop'
+# gem 'jira-ruby', :git => 'https://github.com/sumoheavy/jira-ruby.git'
+# gem 'highline'
+
+# gem 'hashie'
+# gem 'pry'
