+++ Rakefile
@@ -1,0 +1,2 @@
+require "bundler/gem_tasks"
+
